make shared library loader optional
