+++ CMakeLists.txt
@@ -41,6 +41,7 @@ option(ENABLE_EMOJI "Enable emoji module" On)
 option(ENABLE_LIBUUID "Use libuuid for uuid generation" On)
 option(BUILD_SPELL_DICT "Build en_dict.fscd for English spell check" On)
 option(BUILD_SHARED_LIBS "Build library as shared libs" On)
+option(ENABLE_DL "Enable dynamic loading addons" On)
 option(USE_SYSTEM_YOGA "Use system installed yoga library" Off)
 option(USE_SYSTEM_PLASMA_WAYLAND_PROTOCOLS "Use system installed plasma protocols" On)
 set(NO_PREEDIT_APPS "wps.*,wpp.*,et.*" CACHE STRING "Disable preedit for following app by default.")
@@ -120,7 +121,10 @@ endif()
 if(${CMAKE_SYSTEM_NAME} MATCHES "BSD|DragonFly")
     find_package(LibKVM REQUIRED)
 endif()
-find_package(DL REQUIRED)
+
+if (ENABLE_DL)
+    find_package(DL REQUIRED)
+endif()
 
 if (NOT TARGET LibIntl::LibIntl)
     find_package(LibIntl REQUIRED)
