+++ src/lib/fcitx/addonmanager.cpp
@@ -30,6 +30,10 @@
 #include "config.h"
 #include "instance.h"
 
+#ifndef FCITX_NO_DL
+#include "sharedlibraryloader_p.h"
+#endif
+
 namespace fcitx {
 
 class Addon {
@@ -249,7 +253,9 @@ void AddonManager::unregisterLoader(const std::string &name) {
 }
 
 void AddonManager::registerDefaultLoader(StaticAddonRegistry *registry) {
+#ifndef FCITX_NO_DL
     registerLoader(std::make_unique<SharedLibraryLoader>());
+#endif
     if (registry) {
         registerLoader(std::make_unique<StaticLibraryLoader>(registry));
     }
